fix trivial review comments
