+++ node/collation-generation/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 derive_more = "0.99.9"
 futures = "0.3.5"
 log = "0.4.8"
+polkadot-node-primitives = { path = "../primitives" }
 polkadot-node-subsystem = { path = "../subsystem" }
 polkadot-node-subsystem-util = { path = "../subsystem-util" }
 polkadot-primitives = { path = "../../primitives" }
